+++ transport/src/test/java/mobi/nowtechnologies/server/transport/controller/GetChartControllerIT.java
@@ -48,7 +48,7 @@ public class GetChartControllerIT extends AbstractControllerTestIT {
         MediaFile audioMediaFile = mediaFileRepository.save(new MediaFile().withFileName("file name").withFileType(new FileType().withI(MOBILE_AUDIO.getIdAsByte()).withName(MOBILE_AUDIO.name())));
         MediaFile imageFileSmallMediaFile = mediaFileRepository.save(new MediaFile().withFileName("file name").withFileType(new FileType().withI(IMAGE.getIdAsByte()).withName(IMAGE.name())));
         MediaFile imageFileLargeMediaFile = mediaFileRepository.save(new MediaFile().withFileName("file name").withFileType(new FileType().withI(IMAGE.getIdAsByte()).withName(IMAGE.name())));
-        Label label = labelRepository.findOne(1);
+        Label label = labelRepository.findOne((byte)1);
 
         final String isrc = "isrc";
         final long trackId = 666L;
